+++ projects/materialgrammi/src/lib/components/accordion/accordion.component.ts
@@ -7,6 +7,7 @@ import { Theme } from '../../types';
   styleUrls: ['./accordion.component.css']
 })
 export class AccordionComponent implements OnInit {
+  @Input() id = "";
   @Input() theme: Theme = "primary";
 
   @Input() items: any = [{