+++ release-notes.md
@@ -1,3 +1,11 @@
+# Upcoming Features
+- hoscroll
+    - users can create component that can scroll horizontally
+    - have fadded effect on left and right
+    - overflow is hidden
+- badge
+- inview
+
 # ![version 0.1.4](https://img.shields.io/badge/version-1.4-green)
 > 09 April 2022
 
@@ -5,12 +13,17 @@ materialgrammi - ![version 2.0.9](https://img.shields.io/badge/version-2.0.9-blu
 ng-materialgrammi - ![version 0.1.4](https://img.shields.io/badge/version-0.1.4-blue)
 
 ## New Features
-- No
+- Multi Select
+    - multiple values selector
+    - duplicates allow / not-allow option
 
 ## Bug Fixes
 - No
 
 ## Improvement
+- Navbar
+    - added min height (45px) to main container
+    - reduced the space-around from 60 to 50
 - Image
     - added lazy loading and placeholder effect to take place on a new input param